+++ tests/rules/test_transform.py
@@ -43,7 +43,7 @@ from mex.common.types import (
     Year,
     YearMonthDayTime,
 )
-from mex.editor.models import EditorValue
+from mex.editor.models import LANGUAGE_VALUE_NONE, EditorValue
 from mex.editor.rules.models import (
     EditorField,
     EditorPrimarySource,
@@ -182,8 +182,7 @@ def test_get_primary_source_id_from_model_error() -> None:
                     enabled=False,
                 ),
                 EditorValue(
-                    href="http://pavyzdys",
-                    external=True,
+                    href="http://pavyzdys", external=True, badge=LANGUAGE_VALUE_NONE
                 ),
             ],
         ),
@@ -270,7 +269,7 @@ def test_transform_model_values_to_editor_values(
             "AdditiveResource",
             "documentation",
             InputConfig(
-                badge_options=["DE", "EN"],
+                badge_options=["DE", "EN", LANGUAGE_VALUE_NONE],
                 badge_default="DE",
                 badge_titles=["LinkLanguage"],
                 editable_href=True,
@@ -284,7 +283,7 @@ def test_transform_model_values_to_editor_values(
             "AdditiveResource",
             "keyword",
             InputConfig(
-                badge_options=["DE", "EN"],
+                badge_options=["DE", "EN", LANGUAGE_VALUE_NONE],
                 badge_default="DE",
                 badge_titles=["TextLanguage"],
                 editable_badge=True,
@@ -678,6 +677,12 @@ def test_transform_fields_to_preventive(
             Text(language=TextLanguage.DE, value="Beispiel Text"),
         ),
         (
+            EditorValue(text="Text", badge=LANGUAGE_VALUE_NONE),
+            "alternativeTitle",
+            "AdditivePrimarySource",
+            Text(language=None, value="Text"),
+        ),
+        (
             EditorValue(text="ConsentStatus", badge="EXPRESSED_CONSENT"),
             "hasConsentType",
             "AdditiveConsent",
@@ -717,6 +722,7 @@ def test_transform_fields_to_preventive(
     ids=[
         "link",
         "text",
+        "textNoneLang",
         "vocab",
         "default_vocab",
         "temporal",