+++ tests/search/test_transform.py
@@ -1,5 +1,6 @@
 from mex.common.models import PreviewOrganizationalUnit
 from mex.common.types import Text, TextLanguage
+from mex.editor.models import LANGUAGE_VALUE_NONE
 from mex.editor.search.transform import transform_models_to_results
 
 
@@ -20,7 +21,7 @@ def test_transform_models_to_results() -> None:
         "title": [
             {
                 "text": "OU1",
-                "badge": None,
+                "badge": LANGUAGE_VALUE_NONE,
                 "being_edited": False,
                 "href": None,
                 "identifier": None,