+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - infobox to ingest page for "ldap" and "wikidata"
 - new type for aux_provider instead of str
 
+- 'None' as language selection for `Link` and `Text`
+
 ### Changes
 - type of aux provider from `str` to `AuxProvider`